+++ lib/country_flags.dart
@@ -1,7 +1,9 @@
 library country_flags_world;
 
 export 'src/country_flag.dart';
+export 'src/country_flag_with_name.dart';
 export 'src/country_helper.dart';
 export 'src/flag_size.dart';
+export 'src/flag_shape.dart';
 export 'src/shimmer_widget.dart';